feat : write contour

+++ sound-data-converter/src/main/java/com/gaia3d/sound/dataStructure/radiowave/ModelMagneticFieldContour.java
@@ -23,8 +23,8 @@ public class ModelMagneticFieldContour {
     */
 
     private int index;
-    private int x;
-    private int y;
+    private float x;
+    private float y;
     private float relativeHeight;
     private float value;
